Create GEE Script: dNDVI (fast)
Creates Geotiff images for export: 
1. Pre-event
2. Post-event
3. Difference

This fast version does not filter clouds directly, it makes a cloud-free composite by creating a mosaic based on the 'greenest pixel', or the pixels with the maximum NDVI value. 
The dNDVI over land is pretty good and useful for landslide mapping, however, there are strange results over water and the RGB images look a bit strange in places when zoomed in. 

The slow version gives better RGB images.